Breaking: delete AdaptFrameworkUtils, complete utility extraction
Extracts remaining functions from AdaptFrameworkUtils into individual
utility files and removes the class entirely.
