+++ index.js
@@ -5,4 +5,3 @@
 export { default } from './lib/AdaptFrameworkModule.js'
 export { default as AdaptFrameworkBuild } from './lib/AdaptFrameworkBuild.js'
 export { default as AdaptFrameworkImport } from './lib/AdaptFrameworkImport.js'
-export { default as AdaptFrameworkUtils } from './lib/AdaptFrameworkUtils.js'
