+++ lib/AdaptFrameworkBuild.js
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import { App, Hook, ensureDir, writeJson } from 'adapt-authoring-core'
 import { createWriteStream } from 'fs'
 import AdaptCli from 'adapt-cli'
-import FWUtils from './AdaptFrameworkUtils.js'
+import { log, logDir, logMemory, copyFrameworkSource } from './utils.js'
 import fs from 'fs/promises'
 import path from 'upath'
 import semver from 'semver'
@@ -175,15 +175,15 @@ class AdaptFrameworkBuild {
     await ensureDir(this.buildDir)
     await ensureDir(cacheDir)
 
-    FWUtils.logDir('dir', this.dir)
-    FWUtils.logDir('buildDir', this.buildDir)
-    FWUtils.logDir('cacheDir', this.cacheDir)
+    logDir('dir', this.dir)
+    logDir('buildDir', this.buildDir)
+    logDir('cacheDir', this.cacheDir)
 
     await this.loadCourseData()
 
     await Promise.all([
       this.copyAssets(),
-      FWUtils.copyFrameworkSource({
+      copyFrameworkSource({
         destDir: this.dir,
         enabledPlugins: this.enabledPlugins.map(p => p.name),
         linkNodeModules: !this.isExport
@@ -193,11 +193,11 @@ class AdaptFrameworkBuild {
 
     await this.writeContentJson()
 
-    FWUtils.logDir('courseDir', this.courseDir)
+    logDir('courseDir', this.courseDir)
 
     if (!this.isExport) {
       try {
-        FWUtils.logMemory()
+        logMemory()
         await AdaptCli.buildCourse({
           cwd: this.dir,
           sourceMaps: !this.isPublish,
@@ -205,9 +205,9 @@ class AdaptFrameworkBuild {
           cachePath: path.resolve(cacheDir, this.courseId),
           logger: { log: (...args) => App.instance.logger.log('debug', 'adapt-cli', ...args) }
         })
-        FWUtils.logMemory()
+        logMemory()
       } catch (e) {
-        FWUtils.logMemory()
+        logMemory()
         throw App.instance.errors.FW_CLI_BUILD_FAILED
           .setData(e)
       }
@@ -426,7 +426,7 @@ class AdaptFrameworkBuild {
       await ensureDir(dir)
       const filepath = path.join(dir, fileName)
       const returnData = await writeJson(filepath, data)
-      FWUtils.log('verbose', 'WRITE', filepath)
+      log('verbose', 'WRITE', filepath)
       return returnData
     }))
   }
