+++ lib/AdaptFrameworkImport.js
@@ -6,8 +6,7 @@ import path from 'upath'
 import { randomBytes } from 'node:crypto'
 import semver from 'semver'
 import { unzip } from 'zipper'
-import FWUtils from './AdaptFrameworkUtils.js'
-import { getImportContentCounts } from './utils.js'
+import { log, logDir, getImportSummary, getImportContentCounts } from './utils.js'
 
 import ComponentTransform from './migrations/component.js'
 import ConfigTransform from './migrations/config.js'
@@ -256,8 +255,8 @@ class AdaptFrameworkImport {
        */
       this.jsonschema = jsonschema
 
-      FWUtils.log('debug', 'IMPORT_USER', this.userId)
-      FWUtils.log('debug', 'IMPORT_SETTINGS', JSON.stringify(this.settings, null, 2))
+      log('debug', 'IMPORT_USER', this.userId)
+      log('debug', 'IMPORT_SETTINGS', JSON.stringify(this.settings, null, 2))
 
       const { isDryRun, importContent, importPlugins, migrateContent } = this.settings
       const tasks = [
@@ -324,8 +323,8 @@ class AdaptFrameworkImport {
       this.framework.log('error', e)
       throw (e?.statusCode ? e : App.instance.errors.FW_IMPORT_INVALID_COURSE.setData({ reason: e.message }))
     }
-    FWUtils.logDir('unzipPath', this.path)
-    FWUtils.logDir('coursePath', this.coursePath)
+    logDir('unzipPath', this.path)
+    logDir('coursePath', this.coursePath)
 
     try {
       /** @ignore */this.pkg = await readJson(`${this.path}/package.json`)
@@ -345,7 +344,7 @@ class AdaptFrameworkImport {
       this.statusReport.info.push({ code: 'MIGRATE_CONTENT', data })
     }
     await this.convertSchemas()
-    FWUtils.log('debug', 'preparation tasks completed successfully')
+    log('debug', 'preparation tasks completed successfully')
   }
 
   /**
@@ -355,7 +354,7 @@ class AdaptFrameworkImport {
   async convertSchemas () {
     return octopus.runRecursive({
       cwd: this.path,
-      logger: { log: (...args) => FWUtils.log('debug', ...args) }
+      logger: { log: (...args) => log('debug', ...args) }
     })
   }
 
@@ -372,9 +371,9 @@ class AdaptFrameworkImport {
       const customStyle = await fs.readFile(customStylePath, 'utf8')
       const courseJson = await readJson(courseJsonPath)
       await writeJson(courseJsonPath, { customStyle, ...courseJson })
-      FWUtils.log('info', 'patched course customStyle')
+      log('info', 'patched course customStyle')
     } catch (e) {
-      FWUtils.log('warn', 'failed to patch course customStyle', e)
+      log('warn', 'failed to patch course customStyle', e)
     }
   }
 
@@ -388,9 +387,9 @@ class AdaptFrameworkImport {
       if (configJson._theme) return
       configJson._theme = Object.values(this.usedContentPlugins).find(p => p.type === 'theme').name
       await writeJson(configJsonPath, configJson)
-      FWUtils.log('info', 'patched config _theme')
+      log('info', 'patched config _theme')
     } catch (e) {
-      FWUtils.log('warn', 'failed to patch config _theme', e)
+      log('warn', 'failed to patch config _theme', e)
     }
   }
 
@@ -402,14 +401,14 @@ class AdaptFrameworkImport {
     this.assetData = []
     const metaFiles = await glob(`${this.langPath}/assets.json`, { absolute: true })
     if (metaFiles.length) { // process included asset metadata
-      FWUtils.log('debug', 'processing metadata files', metaFiles)
+      log('debug', 'processing metadata files', metaFiles)
       await Promise.all(metaFiles.map(async f => {
         const metaJson = await readJson(f)
         Object.entries(metaJson).forEach(([filename, metadata]) => this.assetData.push({ filename, ...metadata }))
       }))
     } else { // process the file metadata manually
       const assetFiles = await glob(`${this.langPath}/*/*`, { absolute: true })
-      FWUtils.log('debug', 'processing asset files manually', assetFiles.length)
+      log('debug', 'processing asset files manually', assetFiles.length)
       this.assetData.push(...assetFiles.map(f => Object.assign({}, { title: path.basename(f), filepath: f })))
     }
     const hasGlobalTags = !!this.tags.length
@@ -434,7 +433,7 @@ class AdaptFrameworkImport {
     await Promise.all(usedPluginPaths.map(async p => {
       const bowerJson = await readJson(`${p}/bower.json`)
       const { name, version, targetAttribute } = bowerJson
-      FWUtils.log('debug', 'found plugin', name)
+      log('debug', 'found plugin', name)
       this.usedContentPlugins[path.basename(p)] = { name, path: p, version, targetAttribute, type: getPluginType(bowerJson) }
     }))
     this.contentJson.config._enabledPlugins = Object.keys(this.usedContentPlugins)
@@ -448,7 +447,7 @@ class AdaptFrameworkImport {
     const files = await glob('**/*.json', { cwd: this.langPath, absolute: true, ignore: { ignored: p => p.name === 'assets.json' } })
     const mapped = await Promise.all(files.map(f => this.loadContentFile(f)))
     this.statusReport.info.push({ code: 'CONTENT_IMPORTED', data: getImportContentCounts(this.contentJson) })
-    FWUtils.log('info', 'loaded course data successfully')
+    log('info', 'loaded course data successfully')
     return mapped
   }
 
@@ -482,12 +481,12 @@ class AdaptFrameworkImport {
       contents.forEach(c => {
         this.contentJson.contentObjects[c._id] = c
         if (!c._type) {
-          FWUtils.log('warn', App.instance.errors.FW_IMPORT_INVALID_CONTENT.setData({ item: c }))
+          log('warn', App.instance.errors.FW_IMPORT_INVALID_CONTENT.setData({ item: c }))
           this.statusReport.warn.push({ code: 'INVALID_CONTENT', data: c })
         }
       })
     }
-    FWUtils.log('debug', 'LOAD_CONTENT', path.resolve(filePath))
+    log('debug', 'LOAD_CONTENT', path.resolve(filePath))
   }
 
   /**
@@ -518,16 +517,16 @@ class AdaptFrameworkImport {
         captureDir: path.join(`./${migrationId}-migrations`),
         outputFilePath: path.join(this.framework.path, 'migrations', `${migrationId}.txt`)
       }
-      FWUtils.log('debug', 'MIGRATION_ID', migrationId)
-      FWUtils.logDir('captureDir', opts.captureDir)
-      FWUtils.logDir('outputDir', opts.outputDir)
+      log('debug', 'MIGRATION_ID', migrationId)
+      logDir('captureDir', opts.captureDir)
+      logDir('outputDir', opts.outputDir)
 
       await this.runGruntMigration('capture', opts)
       await this.runGruntMigration('migrate', opts)
 
       await fs.rm(path.join(this.framework.path, opts.captureDir), { recursive: true })
     } catch (error) {
-      FWUtils.log('error', 'Migration process failed', error)
+      log('error', 'Migration process failed', error)
       throw App.instance.errors.FW_IMPORT_MIGRATION_FAILED.setData({ reason: error.message })
     }
   }
@@ -569,7 +568,7 @@ class AdaptFrameworkImport {
     if (course.tags) {
       course.tags = course.tags.map(t => existingTagMap[t])
     }
-    FWUtils.log('debug', 'imported tags successfully')
+    log('debug', 'imported tags successfully')
   }
 
   /**
@@ -603,7 +602,7 @@ class AdaptFrameworkImport {
       }
       imagesImported++
     }))
-    FWUtils.log('debug', 'imported course assets successfully')
+    log('debug', 'imported course assets successfully')
     this.statusReport.info.push({ code: 'ASSETS_IMPORTED_SUCCESSFULLY', data: { count: imagesImported } })
   }
 
@@ -633,7 +632,7 @@ class AdaptFrameworkImport {
       const { version: installedVersion, isLocalInstall } = installedP
       if (semver.lt(importVersion, installedVersion)) {
         this.statusReport.info.push({ code: 'PLUGIN_INSTALL_MIGRATING', data: { name: p, installedVersion, importVersion } })
-        FWUtils.log('debug', `migrating '${p}@${importVersion}' during import, installed version is newer (${installedVersion})`)
+        log('debug', `migrating '${p}@${importVersion}' during import, installed version is newer (${installedVersion})`)
         this.pluginsToMigrate.push(p)
         return
       }
@@ -643,12 +642,12 @@ class AdaptFrameworkImport {
       }
       if (semver.eq(importVersion, installedVersion)) {
         this.statusReport.info.push({ code: 'PLUGIN_INSTALL_NOT_NEWER', data: { name: p, installedVersion, importVersion } })
-        FWUtils.log('debug', `not updating '${p}@${importVersion}' during import, installed version is equal to (${installedVersion})`)
+        log('debug', `not updating '${p}@${importVersion}' during import, installed version is equal to (${installedVersion})`)
         return
       }
       if (!isLocalInstall) {
         this.statusReport.warn.push({ code: 'MANAGED_PLUGIN_INSTALL_SKIPPED', data: { name: p, installedVersion, importVersion } })
-        FWUtils.log('debug', `cannot update '${p}' during import, plugin managed via UI`)
+        log('debug', `cannot update '${p}' during import, plugin managed via UI`)
       }
       pluginsToUpdate.push(p)
     })
@@ -685,9 +684,9 @@ class AdaptFrameworkImport {
           this.statusReport.info.push({ code: 'INSTALL_PLUGIN', data: { name: p, version: bowerJson.version } })
         } catch (e) {
           if (e.code === 'EEXIST') {
-            FWUtils.log('warn', 'PLUGIN_ALREADY_EXISTS', p)
+            log('warn', 'PLUGIN_ALREADY_EXISTS', p)
           } else {
-            FWUtils.log('error', 'PLUGIN_IMPORT_FAILED', p, e)
+            log('error', 'PLUGIN_IMPORT_FAILED', p, e)
             errors.push({ plugin: p, error: e.data?.errors?.[0] ?? e })
           }
         }
@@ -700,7 +699,7 @@ class AdaptFrameworkImport {
     this.componentNameMap = Object.values({ ...this.installedPlugins, ...this.newContentPlugins }).reduce((m, v) => {
       return { ...m, [v.targetAttribute.slice(1)]: v.name }
     }, {})
-    FWUtils.log('debug', 'imported course plugins successfully')
+    log('debug', 'imported course plugins successfully')
   }
 
   /**
@@ -745,7 +744,7 @@ class AdaptFrameworkImport {
       }
     }
     if (errors.length) throw App.instance.errors.FW_IMPORT_CONTENT_FAILED.setData({ errors })
-    FWUtils.log('debug', 'imported course data successfully')
+    log('debug', 'imported course data successfully')
   }
 
   /**
@@ -791,7 +790,7 @@ class AdaptFrameworkImport {
     try {
       this.extractAssets(schema.built.properties, insertData)
     } catch (e) {
-      FWUtils.log('error', `failed to extract asset data for attribute '${e.attribute}' of schema '${schemaName}', ${e}`)
+      log('error', `failed to extract asset data for attribute '${e.attribute}' of schema '${schemaName}', ${e}`)
     }
     insertData = await schema.sanitise(insertData)
     let doc
@@ -848,7 +847,7 @@ class AdaptFrameworkImport {
    * @return {Promise}
    */
   async generateSummary () {
-    this.summary = await FWUtils.getImportSummary(this)
+    this.summary = await getImportSummary(this)
   }
 
   /**
@@ -903,10 +902,10 @@ class AdaptFrameworkImport {
 
     const restoreTasks = []
     for (const [pluginName, originalMetadata] of Object.entries(this.updatedContentPlugins)) {
-      FWUtils.log('info', `restoring plugin '${pluginName}' to previous version ${originalMetadata.version}`)
+      log('info', `restoring plugin '${pluginName}' to previous version ${originalMetadata.version}`)
       restoreTasks.push(
         this.contentplugin.restorePluginFromBackup(pluginName)
-          .catch(e => FWUtils.log('error', `failed to restore plugin '${pluginName}' from backup, ${e.message}`))
+          .catch(e => log('error', `failed to restore plugin '${pluginName}' from backup, ${e.message}`))
       )
     }
     return Promise.allSettled(restoreTasks)
