+++ lib/AdaptFrameworkModule.js
@@ -3,7 +3,8 @@ import AdaptFrameworkBuild from './AdaptFrameworkBuild.js'
 import AdaptFrameworkImport from './AdaptFrameworkImport.js'
 import ApiDefs from './apidefs.js'
 import fs from 'fs/promises'
-import FWUtils from './AdaptFrameworkUtils.js'
+import { getHandler, postHandler, importHandler, postUpdateHandler, getUpdateHandler } from './handlers.js'
+import { runCliCommand } from './utils.js'
 import path from 'path'
 import semver from 'semver'
 
@@ -73,10 +74,10 @@ class AdaptFrameworkModule extends AbstractModule {
   }
 
   /**
-   * Reference to AdaptFrameworkUtils#runCliCommand
+   * Reference to runCliCommand utility
    */
   get runCliCommand () {
-    return FWUtils.runCliCommand
+    return runCliCommand
   }
 
   /**
@@ -220,7 +221,7 @@ class AdaptFrameworkModule extends AbstractModule {
       route: '/preview/:id/{*splat}',
       handlers: {
         get: (req, res, next) => { // fail silently
-          FWUtils.getHandler(req, res, e => e ? res.status(e.statusCode || 500).end() : next())
+          getHandler(req, res, e => e ? res.status(e.statusCode || 500).end() : next())
         }
       }
     })
@@ -232,22 +233,22 @@ class AdaptFrameworkModule extends AbstractModule {
     this.apiRouter.addRoute(
       {
         route: '/preview/:id',
-        handlers: { post: FWUtils.postHandler },
+        handlers: { post: postHandler },
         meta: ApiDefs.preview
       },
       {
         route: '/publish/:id',
-        handlers: { post: FWUtils.postHandler, get: FWUtils.getHandler },
+        handlers: { post: postHandler, get: getHandler },
         meta: ApiDefs.publish
       },
       {
         route: '/import',
-        handlers: { post: [FWUtils.importHandler] },
+        handlers: { post: [importHandler] },
         meta: ApiDefs.import
       },
       {
         route: '/export/:id',
-        handlers: { post: FWUtils.postHandler, get: FWUtils.getHandler },
+        handlers: { post: postHandler, get: getHandler },
         meta: ApiDefs.export
       }
     )
@@ -262,7 +263,7 @@ class AdaptFrameworkModule extends AbstractModule {
     if (this.getConfig('enableUpdateApi')) {
       this.apiRouter.addRoute({
         route: '/update',
-        handlers: { post: FWUtils.postUpdateHandler, get: FWUtils.getUpdateHandler },
+        handlers: { post: postUpdateHandler, get: getUpdateHandler },
         meta: ApiDefs.update
       })
       auth.secureRoute(`${this.apiRouter.path}/update`, 'get', ['update:adapt'])
