+++ lib/utils.js
@@ -1,3 +1,9 @@
 export { inferBuildAction } from './utils/inferBuildAction.js'
 export { getPluginUpdateStatus } from './utils/getPluginUpdateStatus.js'
 export { getImportContentCounts } from './utils/getImportContentCounts.js'
+export { log, logDir, logMemory } from './utils/log.js'
+export { runCliCommand } from './utils/runCliCommand.js'
+export { retrieveBuildData } from './utils/retrieveBuildData.js'
+export { getImportSummary } from './utils/getImportSummary.js'
+export { slugifyTitle } from './utils/slugifyTitle.js'
+export { copyFrameworkSource } from './utils/copyFrameworkSource.js'
